Refactoring MasterViewController to Main
