+++ App/TestAppD2/Helpers/Typealias/Closures+Typealias.swift
@@ -1,0 +1,11 @@
+//
+//  Closures.swift
+//  TestAppD2
+//
+//  Created by Arthur Narimanov on 8/7/21.
+//  Copyright © 2021 Григорий Соловьев. All rights reserved.
+//
+
+import Foundation
+
+typealias VoidClosure = () -> Void
